forensics/telemax/src/requirements.txt
pycryptodomex
pypiwin32
requests
pysqlite3